added pagination

+++ 07_lesson/my_project/controllers/exit.php
@@ -4,13 +4,20 @@ if (session_status() === PHP_SESSION_NONE) {
 }
 
 require_once __DIR__ . '/../database/database_connection.php';
-require_once __DIR__ . '/../config.php';
+$connect = connect();
+require_once __DIR__ . '/../functions/database.php';
 
+if (checkAuth($connect)) {
+    require_once __DIR__ . '/../config.php';
+    try {
+        $sql = 'DELETE FROM `users_sessions` WHERE `token` = ?';
+        $stmt = $connect->prepare($sql);
+        $stmt->execute([$_COOKIE['auth']]);
+    } catch (PDOException $exception) {
+        echo $exception->getMessage();
+    }
 
-$sql = 'DELETE FROM `users_sessions` WHERE `token` = ?';
-$stmt = $connect->prepare($sql);
-$stmt->execute([$_COOKIE['auth']]);
+    setcookie('auth', $_COOKIE['auth'], time() - 3600, '/');
 
-setcookie('auth', $_COOKIE['auth'], time() -3600, '/');
-
-header('Location: ' . HOME_PAGE);
+    header('Location: ' . HOME_PAGE);
+}
